+++ src/main/java/rs117/hd/HdPlugin.java
@@ -108,7 +108,10 @@ import rs117.hd.environments.EnvironmentManager;
 import rs117.hd.lighting.LightManager;
 import rs117.hd.lighting.SceneLight;
 import rs117.hd.materials.Material;
-import rs117.hd.materials.ObjectProperties;
+import rs117.hd.model.objects.ObjectProperties;
+import rs117.hd.model.ModelHasher;
+import rs117.hd.model.ModelPusher;
+import rs117.hd.model.objects.ObjectType;
 import rs117.hd.opengl.compute.OpenCLManager;
 import rs117.hd.opengl.shader.Shader;
 import rs117.hd.opengl.shader.ShaderException;
