+++ src/main/java/rs117/hd/SceneUploader.java
@@ -44,10 +44,12 @@ import net.runelite.api.SceneTilePaint;
 import net.runelite.api.Tile;
 import net.runelite.api.WallObject;
 import rs117.hd.materials.GroundMaterial;
-import rs117.hd.materials.ObjectProperties;
+import rs117.hd.model.objects.ObjectProperties;
 import rs117.hd.materials.Overlay;
 import rs117.hd.materials.Material;
 import rs117.hd.materials.Underlay;
+import rs117.hd.model.ModelPusher;
+import rs117.hd.model.objects.ObjectType;
 import rs117.hd.utils.HDUtils;
 import rs117.hd.utils.buffer.GpuFloatBuffer;
 import rs117.hd.utils.buffer.GpuIntBuffer;
